requirements.txt
asgiref==3.2.9
Django==3.0.7
djangorestframework==3.11.0
Pillow==7.1.2
pytz==2020.1
six==1.15.0
slugify==0.0.1
sqlparse==0.3.1
transliterate==1.10.2
